% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{chat}
\alias{chat}
\alias{initiate_llm}
\alias{new_chat}
\alias{chat_in_browser}
\alias{chat_in_console}
\alias{get_provider}
\alias{get_system_prompt}
\alias{get_tokens}
\title{Chat With Local LLM}
\usage{
initiate_llm(
  address = read_secret("llm.address"),
  port = read_secret("llm.port"),
  system_prompt = read_secret("llm.system_prompt"),
  model = read_secret("llm.model"),
  ...
)

new_chat(input, ...)

chat(input, new = FALSE, ...)

chat_in_browser(new = FALSE, ...)

chat_in_console(new = FALSE, ...)

get_provider()

get_system_prompt()

get_tokens()
}
\arguments{
\item{address}{Server address, defaults to local.}

\item{port}{Server port, defaults to local.}

\item{system_prompt}{A system prompt to set the behavior of the assistant.}

\item{model}{The model to use for the chat.
Use \code{models_ollama()} to see all options.}

\item{...}{Argument passed on to \code{\link[ellmer:chat_ollama]{ellmer::chat_ollama()}}.}

\item{input}{Text for input.}

\item{new}{Initiate new LLM instance.}
}
\description{
Chat With Local LLM
}
