% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{chat}
\alias{chat}
\alias{initiate_llm}
\alias{chat_in_browser}
\alias{chat_in_console}
\alias{get_provider}
\alias{get_system_prompt}
\alias{get_tokens}
\title{Chat With Local LLM}
\usage{
initiate_llm(
  address = "http://localhost",
  port = 11434,
  system_prompt = read_secret("llm.system_prompt"),
  model = read_secret("llm.default.model"),
  ...
)

chat(input, ...)

chat_in_browser(...)

chat_in_console(...)

get_provider()

get_system_prompt()

get_tokens()
}
\arguments{
\item{address}{Server address, defaults to local.}

\item{port}{Server port, defaults to local.}

\item{system_prompt}{A system prompt to set the behavior of the assistant.}

\item{model}{The model to use for the chat.
Use \code{models_ollama()} to see all options.}

\item{...}{Argument passed on to \code{\link[ellmer:chat_ollama]{ellmer::chat_ollama()}}.}

\item{input}{Text for input.}
}
\description{
Chat With Local LLM
}
